requirements.txt
Django==5.2.1
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
pillow==11.2.1
pip==23.2.1
python-dotenv==1.1.0
PyYAML==6.0.2